Improve robustness of longlived_name migration
On InfraQA, this migration is failing. The reason is that there are
invalid entries in the search_config table. These invalid entries
are referring to tables in GenericDB that no longer exist.

Now, as a final cleanup step, we remove all such invalid rows.

As part of this work, I also made a simplifying change, which is to
simply query ALL tables in the GenericDB, instead of limiting it
to the set of tables in the search config.