RowKey is now returned with the search results (#4)
* RowKey is now returned with the search results
RowKey either defaults to `rowid` or is replaced with the indexkey specified in
the config. The idea is that the requesting client can now use this key to
query crud for the actual row. Also moved all packages to `server`... it is a
bad name, and maybe some of the files should live in a package call server but
not all. This was done as compiling this project on linux creates an executable
called search which matched the search folder :( also changed the doc.go file to
doc.md and added some md styling. TI-3536

* Removed `ItemType` as it does not add value.