INF-1081: IMQS search unresponsive (#6)
This is my attempt to do a few things:
- reinvent some of the search service conf so that we no longer have to
specify database connection details, but rather use db-aliases.json that
comes from most of the confs that we provide. This change seemed all the
more obvious considering that the db-alias name was already used to
identify the field information to which these
- Add auto vacuuming to the search service (making sure that it happens
at 2am every day)
- If the search service is running in Windows AT ALL, then there will be
no access to logging to stdout. I reckon this is a safer assumption
- In some other services that make use of the log dependency have the
disable std-out option set on the service-registry.json file
- However, there are only a vanishingly small number of developers who
still work on Windows machines, and furthermore, the ones who still work
on Windows have almost no reason to be looking for stdout logging from
the search-service and considering our production machines ought to log
to files, rather than adding the option to the service-registry (a 
customization option that adds no real value, I chose to rather discontinue
supporting logging to stdout when services are running in Windows.